app/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==20.1.0
python-dotenv==0.19.0
psycopg2-binary==2.9.3
requests==2.26.0
Flask-SQLAlchemy==3.1.1
pymysql==1.0.2
cryptography==3.4.8
Pillow==8.3.2
APScheduler==3.7.0
Flask-Caching==1.10.1
Flask-Mail==0.9.1
itsdangerous==2.1.2
python-dateutil==2.8.2
schedule==1.1.0


